+++ internal/services/keepservice.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 
-	"github.com/Dorrrke/GophKeeper-client/internal/client"
 	errText "github.com/Dorrrke/GophKeeper-client/internal/domain/errors"
 	"github.com/Dorrrke/GophKeeper-client/internal/domain/models"
 	"golang.org/x/crypto/bcrypt"
@@ -13,43 +12,74 @@ import (
 var ErrInvalidPassword = errors.New(errText.InvalidPasswordError)
 
 type Storage interface {
-	SaveUser(ctx context.Context, user models.UserModel) (int64, error)
-	GetUserHash(ctx context.Context, login string) (int64, string, error)
-	SaveCard(ctx context.Context, card models.CardModel, uID int64) (int64, error)
-	SaveLogin(ctx context.Context, login models.LoginModel, uID int64) (int64, error)
-	SaveText(ctx context.Context, text models.TextDataModel, uID int64) (int64, error)
-	SaveBin(ctx context.Context, binData models.BinaryDataModel, uID int64) (int64, error)
-	GetAllCards(ctx context.Context, uID int64) ([]models.CardModel, error)
-	GetAllBin(ctx context.Context, uID int64) ([]models.BinaryDataModel, error)
-	GetAllLogins(ctx context.Context, uID int64) ([]models.LoginModel, error)
-	GetAllTextData(ctx context.Context, uID int64) ([]models.TextDataModel, error)
-	GetCardByName(ctx context.Context, name string, uID int64) (models.CardModel, error)
-	GetLoginByName(ctx context.Context, name string, uID int64) (models.LoginModel, error)
-	GetTextDataByName(ctx context.Context, name string, uID int64) (models.TextDataModel, error)
-	GetBinByName(ctx context.Context, name string, uID int64) (models.BinaryDataModel, error)
-	DeleteCard(ctx context.Context, name string, uID int64) error
-	DeleteLogin(ctx context.Context, name string, uID int64) error
-	DeleteText(ctx context.Context, name string, uID int64) error
-	DeleteBin(ctx context.Context, name string, uID int64) error
-	UpdateCard(ctx context.Context, card models.CardModel, uID int64) error
-	UpdateLogin(ctx context.Context, auth models.LoginModel, uID int64) error
-	UpdateText(ctx context.Context, data models.TextDataModel, uID int64) error
-	UpdateBin(ctx context.Context, data models.BinaryDataModel, uID int64) error
-	GetAllSaves(ctx context.Context, uID int64) (models.SyncModel, error)
 	Sync(ctx context.Context, model models.SyncModel) error
 	ClearDB(ctx context.Context, uId int64) error
+	GetAllSaves(ctx context.Context, uID int64) (models.SyncModel, error)
+}
+
+type UserStorage interface {
+	SaveUser(ctx context.Context, user models.UserModel) (int64, error)
+	GetUserHash(ctx context.Context, login string) (int64, string, error)
+}
+
+type CardStorage interface {
+	SaveCard(ctx context.Context, card models.CardModel, uID int64) (int64, error)
+	GetAllCards(ctx context.Context, uID int64) ([]models.CardModel, error)
+	GetCardByName(ctx context.Context, name string, uID int64) (models.CardModel, error)
+	DeleteCard(ctx context.Context, name string, uID int64) error
+	UpdateCard(ctx context.Context, card models.CardModel, uID int64) error
+}
+
+type AuthStorage interface {
+	SaveLogin(ctx context.Context, login models.LoginModel, uID int64) (int64, error)
+	GetAllLogins(ctx context.Context, uID int64) ([]models.LoginModel, error)
+	GetLoginByName(ctx context.Context, name string, uID int64) (models.LoginModel, error)
+	DeleteLogin(ctx context.Context, name string, uID int64) error
+	UpdateLogin(ctx context.Context, auth models.LoginModel, uID int64) error
+}
+
+type BinStorage interface {
+	SaveBin(ctx context.Context, binData models.BinaryDataModel, uID int64) (int64, error)
+	GetAllBin(ctx context.Context, uID int64) ([]models.BinaryDataModel, error)
+	GetBinByName(ctx context.Context, name string, uID int64) (models.BinaryDataModel, error)
+	DeleteBin(ctx context.Context, name string, uID int64) error
+	UpdateBin(ctx context.Context, data models.BinaryDataModel, uID int64) error
+}
+
+type TextStorage interface {
+	SaveText(ctx context.Context, text models.TextDataModel, uID int64) (int64, error)
+	GetAllTextData(ctx context.Context, uID int64) ([]models.TextDataModel, error)
+	GetTextDataByName(ctx context.Context, name string, uID int64) (models.TextDataModel, error)
+	DeleteText(ctx context.Context, name string, uID int64) error
+	UpdateText(ctx context.Context, data models.TextDataModel, uID int64) error
+}
+
+type Client interface {
+	Register(ctx context.Context, login, password string) error
+	Login(ctx context.Context, login, password string) error
+	Sync(ctx context.Context, model models.SyncModel, uID int64) (models.SyncModel, error)
 }
 
 type KeepService struct {
-	keepClient *client.KeeperClient
+	keepClient Client
 	stor       Storage
+	userStor   UserStorage
+	cardStor   CardStorage
+	textStor   TextStorage
+	binStor    BinStorage
+	authStor   AuthStorage
 }
 
 // TODO: Добавить килент
-func New(client *client.KeeperClient, stor Storage) *KeepService {
+func New(client Client, stor Storage, uStor UserStorage, cStor CardStorage, tStor TextStorage, bStor BinStorage, aStor AuthStorage) *KeepService {
 	return &KeepService{
 		keepClient: client,
 		stor:       stor,
+		userStor:   uStor,
+		cardStor:   cStor,
+		textStor:   tStor,
+		binStor:    bStor,
+		authStor:   aStor,
 	}
 }
 
@@ -58,7 +88,7 @@ func (kp *KeepService) RegisterUser(login string, pass string) (int64, error) {
 	if err != nil {
 		return -1, err
 	}
-	uID, err := kp.stor.SaveUser(context.Background(), models.UserModel{
+	uID, err := kp.userStor.SaveUser(context.Background(), models.UserModel{
 		Login: login,
 		Hash:  hash,
 	})
@@ -69,7 +99,7 @@ func (kp *KeepService) RegisterUser(login string, pass string) (int64, error) {
 }
 
 func (kp *KeepService) LoginUser(login string, pass string) (models.UserModel, error) {
-	uID, hashFromDB, err := kp.stor.GetUserHash(context.Background(), login)
+	uID, hashFromDB, err := kp.userStor.GetUserHash(context.Background(), login)
 	if err != nil {
 		return models.UserModel{}, err
 	}
@@ -84,7 +114,7 @@ func (kp *KeepService) LoginUser(login string, pass string) (models.UserModel, e
 }
 
 func (kp *KeepService) SaveCard(card models.CardModel, uID int64) (int64, error) {
-	cID, err := kp.stor.SaveCard(context.Background(), card, uID)
+	cID, err := kp.cardStor.SaveCard(context.Background(), card, uID)
 	if err != nil {
 		return -1, err
 	}
@@ -92,7 +122,7 @@ func (kp *KeepService) SaveCard(card models.CardModel, uID int64) (int64, error)
 }
 
 func (kp *KeepService) SaveLogin(loginData models.LoginModel, uID int64) (int64, error) {
-	lID, err := kp.stor.SaveLogin(context.Background(), loginData, uID)
+	lID, err := kp.authStor.SaveLogin(context.Background(), loginData, uID)
 	if err != nil {
 		return -1, err
 	}
@@ -100,7 +130,7 @@ func (kp *KeepService) SaveLogin(loginData models.LoginModel, uID int64) (int64,
 }
 
 func (kp *KeepService) SaveTextData(textData models.TextDataModel, uID int64) (int64, error) {
-	tID, err := kp.stor.SaveText(context.Background(), textData, uID)
+	tID, err := kp.textStor.SaveText(context.Background(), textData, uID)
 	if err != nil {
 		return -1, err
 	}
@@ -108,7 +138,7 @@ func (kp *KeepService) SaveTextData(textData models.TextDataModel, uID int64) (i
 }
 
 func (kp *KeepService) SaveBinaryData(binData models.BinaryDataModel, uID int64) (int64, error) {
-	bID, err := kp.stor.SaveBin(context.Background(), binData, uID)
+	bID, err := kp.binStor.SaveBin(context.Background(), binData, uID)
 	if err != nil {
 		return -1, err
 	}
@@ -116,7 +146,7 @@ func (kp *KeepService) SaveBinaryData(binData models.BinaryDataModel, uID int64)
 }
 
 func (kp *KeepService) GetBins(uID int64) ([]models.BinaryDataModel, error) {
-	bins, err := kp.stor.GetAllBin(context.Background(), uID)
+	bins, err := kp.binStor.GetAllBin(context.Background(), uID)
 	if err != nil {
 		return nil, err
 	}
@@ -124,7 +154,7 @@ func (kp *KeepService) GetBins(uID int64) ([]models.BinaryDataModel, error) {
 }
 
 func (kp *KeepService) GetCards(uID int64) ([]models.CardModel, error) {
-	cards, err := kp.stor.GetAllCards(context.Background(), uID)
+	cards, err := kp.cardStor.GetAllCards(context.Background(), uID)
 	if err != nil {
 		return nil, err
 	}
@@ -132,7 +162,7 @@ func (kp *KeepService) GetCards(uID int64) ([]models.CardModel, error) {
 }
 
 func (kp *KeepService) GetLogins(uID int64) ([]models.LoginModel, error) {
-	logins, err := kp.stor.GetAllLogins(context.Background(), uID)
+	logins, err := kp.authStor.GetAllLogins(context.Background(), uID)
 	if err != nil {
 		return nil, err
 	}
@@ -140,7 +170,7 @@ func (kp *KeepService) GetLogins(uID int64) ([]models.LoginModel, error) {
 }
 
 func (kp *KeepService) GetTextData(uID int64) ([]models.TextDataModel, error) {
-	tData, err := kp.stor.GetAllTextData(context.Background(), uID)
+	tData, err := kp.textStor.GetAllTextData(context.Background(), uID)
 	if err != nil {
 		return nil, err
 	}
@@ -148,7 +178,7 @@ func (kp *KeepService) GetTextData(uID int64) ([]models.TextDataModel, error) {
 }
 
 func (kp *KeepService) GetCardByName(cName string, uID int64) (models.CardModel, error) {
-	card, err := kp.stor.GetCardByName(context.Background(), cName, uID)
+	card, err := kp.cardStor.GetCardByName(context.Background(), cName, uID)
 	if err != nil {
 		return models.CardModel{}, err
 	}
@@ -156,7 +186,7 @@ func (kp *KeepService) GetCardByName(cName string, uID int64) (models.CardModel,
 }
 
 func (kp *KeepService) GetLoginByName(lName string, uID int64) (models.LoginModel, error) {
-	login, err := kp.stor.GetLoginByName(context.Background(), lName, uID)
+	login, err := kp.authStor.GetLoginByName(context.Background(), lName, uID)
 	if err != nil {
 		return models.LoginModel{}, err
 	}
@@ -164,7 +194,7 @@ func (kp *KeepService) GetLoginByName(lName string, uID int64) (models.LoginMode
 }
 
 func (kp *KeepService) GetTextDataByName(tName string, uID int64) (models.TextDataModel, error) {
-	tData, err := kp.stor.GetTextDataByName(context.Background(), tName, uID)
+	tData, err := kp.textStor.GetTextDataByName(context.Background(), tName, uID)
 	if err != nil {
 		return models.TextDataModel{}, err
 	}
@@ -172,7 +202,7 @@ func (kp *KeepService) GetTextDataByName(tName string, uID int64) (models.TextDa
 }
 
 func (kp *KeepService) GetBinByName(tName string, uID int64) (models.BinaryDataModel, error) {
-	tData, err := kp.stor.GetBinByName(context.Background(), tName, uID)
+	tData, err := kp.binStor.GetBinByName(context.Background(), tName, uID)
 	if err != nil {
 		return models.BinaryDataModel{}, err
 	}
@@ -180,40 +210,40 @@ func (kp *KeepService) GetBinByName(tName string, uID int64) (models.BinaryDataM
 }
 
 func (kp *KeepService) DeleteCardByName(name string, uID int64) error {
-	err := kp.stor.DeleteCard(context.Background(), name, uID)
+	err := kp.cardStor.DeleteCard(context.Background(), name, uID)
 	return err
 }
 
 func (kp *KeepService) DeleteLoginByName(name string, uID int64) error {
-	err := kp.stor.DeleteLogin(context.Background(), name, uID)
+	err := kp.authStor.DeleteLogin(context.Background(), name, uID)
 	return err
 }
 func (kp *KeepService) DeleteTextByName(name string, uID int64) error {
-	err := kp.stor.DeleteText(context.Background(), name, uID)
+	err := kp.textStor.DeleteText(context.Background(), name, uID)
 	return err
 }
 func (kp *KeepService) DeleteBinByName(name string, uID int64) error {
-	err := kp.stor.DeleteBin(context.Background(), name, uID)
+	err := kp.binStor.DeleteBin(context.Background(), name, uID)
 	return err
 }
 
 func (kp *KeepService) UpdateCard(card models.CardModel, uID int64) error {
-	err := kp.stor.UpdateCard(context.Background(), card, uID)
+	err := kp.cardStor.UpdateCard(context.Background(), card, uID)
 	return err
 }
 
 func (kp *KeepService) UpdateLogin(auth models.LoginModel, uID int64) error {
-	err := kp.stor.UpdateLogin(context.Background(), auth, uID)
+	err := kp.authStor.UpdateLogin(context.Background(), auth, uID)
 	return err
 }
 
 func (kp *KeepService) UpdateText(data models.TextDataModel, uID int64) error {
-	err := kp.stor.UpdateText(context.Background(), data, uID)
+	err := kp.textStor.UpdateText(context.Background(), data, uID)
 	return err
 }
 
 func (kp *KeepService) UpdateBin(data models.BinaryDataModel, uID int64) error {
-	err := kp.stor.UpdateBin(context.Background(), data, uID)
+	err := kp.binStor.UpdateBin(context.Background(), data, uID)
 	return err
 }
